@@ -26,26 +26,18 @@ namespace Corvus.CosmosClient.Extensions
         /// <param name="requestOptions">Request options for the query.</param>
         /// <param name="maxBatchCount">The maximum number of batches to process.</param>
         /// <param name="continuationToken">The continuation token from which to resume processing.</param>
+        /// <param name="feedRange">The feed range over which to execute the iterator.</param>
         /// <param name="cancellationToken">A cancellation token to terminate the option early.</param>
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
-        public static Task<string?> ForEachAsync<T>(this Container container, string queryText, Action<T> action, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
+        public static Task<string?> ForEachAsync<T>(this Container container, string queryText, Action<T> action, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, FeedRange? feedRange = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryText is null)
-            {
-                throw new ArgumentNullException(nameof(queryText));
-            }
+            ArgumentNullException.ThrowIfNull(queryText);
 
-            if (action is null)
-            {
-                throw new ArgumentNullException(nameof(action));
-            }
+            ArgumentNullException.ThrowIfNull(action);
 
-            return ForEachAsync(container, new QueryDefinition(queryText), action, requestOptions, maxBatchCount, continuationToken, cancellationToken);
+            return ForEachAsync(container, new QueryDefinition(queryText), action, requestOptions, maxBatchCount, continuationToken, feedRange, cancellationToken);
         }
 
         /// <summary>
@@ -58,26 +50,21 @@ namespace Corvus.CosmosClient.Extensions
         /// <param name="requestOptions">Request options for the query.</param>
         /// <param name="maxBatchCount">The maximum number of batches to process.</param>
         /// <param name="continuationToken">The continuation token from which to resume processing.</param>
+        /// <param name="feedRange">The feed range over which to execute the iterator.</param>
         /// <param name="cancellationToken">A cancellation token to terminate the option early.</param>
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
-        public static async Task<string?> ForEachAsync<T>(this Container container, QueryDefinition queryDefinition, Action<T> action, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
+        public static async Task<string?> ForEachAsync<T>(this Container container, QueryDefinition queryDefinition, Action<T> action, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, FeedRange? feedRange = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryDefinition == null)
-            {
-                throw new ArgumentNullException(nameof(queryDefinition));
-            }
+            ArgumentNullException.ThrowIfNull(queryDefinition);
 
-            if (action == null)
-            {
-                throw new ArgumentNullException(nameof(action));
-            }
+            ArgumentNullException.ThrowIfNull(action);
 
-            FeedIterator<T> iterator = container.GetItemQueryIterator<T>(queryDefinition, continuationToken, requestOptions);
+            FeedIterator<T> iterator =
+                feedRange is FeedRange fr ?
+                    container.GetItemQueryIterator<T>(fr, queryDefinition, continuationToken, requestOptions) :
+                    container.GetItemQueryIterator<T>(queryDefinition, continuationToken, requestOptions);
 
             int batchCount = 0;
             string? previousContinuationToken = null;
@@ -129,26 +116,18 @@ namespace Corvus.CosmosClient.Extensions
         /// <param name="requestOptions">Request options for the query.</param>
         /// <param name="maxBatchCount">The maximum number of batches to process.</param>
         /// <param name="continuationToken">The continuation token from which to resume processing.</param>
+        /// <param name="feedRange">The feed range over which to execute the iterator.</param>
         /// <param name="cancellationToken">A cancellation token to terminate the option early.</param>
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
-        public static Task<string?> ForEachAsync<T>(this Container container, string queryText, Func<T, Task> actionAsync, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
+        public static Task<string?> ForEachAsync<T>(this Container container, string queryText, Func<T, Task> actionAsync, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, FeedRange? feedRange = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryText is null)
-            {
-                throw new ArgumentNullException(nameof(queryText));
-            }
+            ArgumentNullException.ThrowIfNull(queryText);
 
-            if (actionAsync is null)
-            {
-                throw new ArgumentNullException(nameof(actionAsync));
-            }
+            ArgumentNullException.ThrowIfNull(actionAsync);
 
-            return ForEachAsync(container, new QueryDefinition(queryText), actionAsync, requestOptions, maxBatchCount, continuationToken, cancellationToken);
+            return ForEachAsync(container, new QueryDefinition(queryText), actionAsync, requestOptions, maxBatchCount, continuationToken, feedRange, cancellationToken);
         }
 
         /// <summary>
@@ -161,26 +140,20 @@ namespace Corvus.CosmosClient.Extensions
         /// <param name="requestOptions">Request options for the query.</param>
         /// <param name="maxBatchCount">The maximum number of batches to process.</param>
         /// <param name="continuationToken">The continuation token from which to resume processing.</param>
+        /// <param name="feedRange">The feed range over which to execute the iterator.</param>
         /// <param name="cancellationToken">A cancellation token to terminate the option early.</param>
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
-        public static async Task<string?> ForEachAsync<T>(this Container container, QueryDefinition queryDefinition, Func<T, Task> actionAsync, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
+        public static async Task<string?> ForEachAsync<T>(this Container container, QueryDefinition queryDefinition, Func<T, Task> actionAsync, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, FeedRange? feedRange = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryDefinition == null)
-            {
-                throw new ArgumentNullException(nameof(queryDefinition));
-            }
+            ArgumentNullException.ThrowIfNull(queryDefinition);
 
-            if (actionAsync == null)
-            {
-                throw new ArgumentNullException(nameof(actionAsync));
-            }
+            ArgumentNullException.ThrowIfNull(actionAsync);
 
-            FeedIterator<T> iterator = container.GetItemQueryIterator<T>(queryDefinition, continuationToken, requestOptions);
+            FeedIterator<T> iterator = feedRange is FeedRange fr ?
+                container.GetItemQueryIterator<T>(fr, queryDefinition, continuationToken, requestOptions) :
+                container.GetItemQueryIterator<T>(queryDefinition, continuationToken, requestOptions);
 
             int batchCount = 0;
             string? previousContinuationToken = null;
